requirements.txt
python-dotenv==1.0.1
fastapi==0.115.6
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-monitoring==2.24.0
google-cloud-monitoring-dashboards==2.17.0
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.1
google-cloud-monitoring>=2.24.0
torch-geometric==2.6.1
transformers==4.47.0
pandas==2.2.3
numpy==1.26.4
redis==5.2.1
scikit-learn==1.6.0
tqdm==4.66.2
uvicorn==0.34.0
vertexai>=0.0.1  
db-dtypes==1.3.1